move `cross_table_argument.rs` into folder `table`
