+++ triton-vm/src/table/jump_stack_table.rs
@@ -21,8 +21,6 @@ use std::fmt::Display;
 use std::fmt::Formatter;
 use JumpStackTableChallengeId::*;
 
-use crate::cross_table_arguments::CrossTableArg;
-use crate::cross_table_arguments::PermArg;
 use crate::instruction::Instruction;
 use crate::table::challenges::TableChallenges;
 use crate::table::constraint_circuit::ConstraintCircuit;
@@ -31,6 +29,8 @@ use crate::table::constraint_circuit::DualRowIndicator;
 use crate::table::constraint_circuit::DualRowIndicator::*;
 use crate::table::constraint_circuit::SingleRowIndicator;
 use crate::table::constraint_circuit::SingleRowIndicator::*;
+use crate::table::cross_table_argument::CrossTableArg;
+use crate::table::cross_table_argument::PermArg;
 use crate::table::table_collection::NUM_BASE_COLUMNS;
 use crate::table::table_collection::NUM_EXT_COLUMNS;
 use crate::table::table_column::BaseTableColumn;
