+++ triton-vm/src/table/ram_table.rs
@@ -20,8 +20,6 @@ use twenty_first::shared_math::x_field_element::XFieldElement;
 
 use RamTableChallengeId::*;
 
-use crate::cross_table_arguments::CrossTableArg;
-use crate::cross_table_arguments::PermArg;
 use crate::instruction::Instruction;
 use crate::table::challenges::TableChallenges;
 use crate::table::constraint_circuit::ConstraintCircuit;
@@ -30,6 +28,8 @@ use crate::table::constraint_circuit::DualRowIndicator;
 use crate::table::constraint_circuit::DualRowIndicator::*;
 use crate::table::constraint_circuit::SingleRowIndicator;
 use crate::table::constraint_circuit::SingleRowIndicator::*;
+use crate::table::cross_table_argument::CrossTableArg;
+use crate::table::cross_table_argument::PermArg;
 use crate::table::table_collection::NUM_BASE_COLUMNS;
 use crate::table::table_collection::NUM_EXT_COLUMNS;
 use crate::table::table_column::BaseTableColumn;
